feat: :art: adding onetomany

+++ survey/src/main/java/com/proyectospringsurvey/survey/surveysCategory/infrastructure/ImpServicesCategory.java
@@ -5,8 +5,10 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-
+import com.proyectospringsurvey.survey.surveys.domain.Surveys;
+import com.proyectospringsurvey.survey.surveys.infrastructure.SurveyRepository;
 import com.proyectospringsurvey.survey.surveysCategory.application.ISurveyCategory;
 import com.proyectospringsurvey.survey.surveysCategory.domain.surveysCategory;
 
@@ -15,6 +17,9 @@ public class ImpServicesCategory implements ISurveyCategory{
     
     @Autowired
     private RepositorysCategory repositorysCategory;
+
+    @Autowired
+    SurveyRepository surveyRepository;
 
     @Override
     public surveysCategory createsCategory(surveysCategory sCategory) {
@@ -31,9 +36,20 @@ public class ImpServicesCategory implements ISurveyCategory{
         return repositorysCategory.findById(id);
     }
 
+     
     @Override
-    public void DeletesCategoryByName(String name) {
-        repositorysCategory.deleteByName(name);
+    @Transactional
+    public boolean DeletesCategoryByName(String name) {
+        surveysCategory categoria = repositorysCategory.findByname(name);
+        if(categoria.getName()!=name){
+            categoria.getSurveys().removeAll(categoria.getSurveys());
+            repositorysCategory.save(categoria);
+            repositorysCategory.deleteByName(name);
+            return true;
+        }else{
+            return false;
+        }
+
     }
